Extract inner class
