+++ src/main/java/alfio/plugin/PluginDataStorageProvider.java
@@ -52,55 +52,4 @@ public class PluginDataStorageProvider {
     }
 
 
-    public static class PluginDataStorage {
-        private final String pluginId;
-        private final PluginConfigurationRepository pluginConfigurationRepository;
-        private final PluginLogRepository pluginLogRepository;
-        private final TransactionTemplate tx;
-        private final EventRepository eventRepository;
-
-        private PluginDataStorage(String pluginId,
-                                  PluginConfigurationRepository pluginConfigurationRepository,
-                                  PluginLogRepository pluginLogRepository, TransactionTemplate tx, EventRepository eventRepository) {
-            this.pluginId = pluginId;
-            this.pluginConfigurationRepository = pluginConfigurationRepository;
-            this.pluginLogRepository = pluginLogRepository;
-            this.tx = tx;
-            this.eventRepository = eventRepository;
-        }
-
-        public Event getEventById(int eventId) {
-            return eventRepository.findById(eventId);
-        }
-
-        public Optional<String> getConfigValue(String name, int eventId) {
-            return pluginConfigurationRepository.loadSingleOption(pluginId, eventId, name).map(PluginConfigOption::getOptionValue);
-        }
-
-        public void insertConfigValue(int eventId, String name, String value, String description, ComponentType componentType) {
-            pluginConfigurationRepository.insert(pluginId, eventId, name, value, description, componentType);
-        }
-
-        public void registerSuccess(String description, int eventId) {
-            tx.execute(tc -> {
-                pluginLogRepository.insertEvent(pluginId, eventId, description, PluginLog.Type.SUCCESS, ZonedDateTime.now(Clock.systemUTC()));
-                return null;
-            });
-        }
-
-        public void registerFailure(String description, int eventId) {
-            tx.execute(tc -> {
-                pluginLogRepository.insertEvent(pluginId, eventId, description, PluginLog.Type.ERROR, ZonedDateTime.now(Clock.systemUTC()));
-                return null;
-            });
-        }
-
-        public void registerWarning(String description, int eventId) {
-            tx.execute(tc -> {
-                pluginLogRepository.insertEvent(pluginId, eventId, description, PluginLog.Type.WARNING, ZonedDateTime.now(Clock.systemUTC()));
-                return null;
-            });
-        }
-
-    }
 }
